Changes made for Support for Gyroscope Limited Axes sensor #43
